+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## v0.10.0 (TBD)
 
+### Changes
+
+- Migrate from `rusqlite` to `diesel` as a database abstraction (#921)
+
 ## v0.9.0 (2025-05-30)
 
 ### Enhancements